WIP adding tests
